Better types for the entire project.
We move away from shared_ptr in many cases, and we use pass-by-value,
constant references, and standard C++ move semantics to get cleaner
and simpler code.

The Event object has just become an alias for Object (which itself is
an alias for the JSON object); we've provided nice factory functions
for this thing for the various types.  We also offer up a PostEvent
API that lets consumers post events that they have constructed directly.

There are also APIs to allow consumers convenience methods to adjust
things like anonymous ID, integrations, etc. on the top-level object.